requirement.txt
fastapi==0.110.0
uvicorn==0.27.1
playwright==1.42.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
